bbs-project/bbs-project/client-python/requirements.txt
pyzmq==25.1.2
msgpack==1.0.8
